+++ packages/isar_test/test/copy_to_file_test.dart
@@ -112,6 +112,7 @@ void main() {
           '$copiedDbFilename2.isar',
         ),
       );
+      await isarCopy1.copyToFile(copiedDbFile2.path);
 
       expect(copiedDbFile2.lengthSync(), greaterThan(0));
       expect(